# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API is a user-friendly solution designed to serve as the foundation for more complex APIs.",
        "title": "Go API",
        "contact": {
            "name": "Raul del Aguila",
            "url": "https://github.com/raulaguila",
            "email": "email@email.com"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/user": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get users",
                "parameters": [
                    {
                        "enum": [
                            true,
                            false
                        ],
                        "type": "boolean",
                        "default": true,
                        "description": "Skip auth",
                        "name": "X-Skip-Auth",
                        "in": "header"
                    },
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "role_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.PaginatedOutput-github_com_raulaguila_go-api_internal_core_dto_UserOutput"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Insert user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Insert user",
                "parameters": [
                    {
                        "enum": [
                            true,
                            false
                        ],
                        "type": "boolean",
                        "default": true,
                        "description": "Skip auth",
                        "name": "X-Skip-Auth",
                        "in": "header"
                    },
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "description": "User model",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.UserInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.UserOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete user by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Delete user by ID",
                "parameters": [
                    {
                        "enum": [
                            true,
                            false
                        ],
                        "type": "boolean",
                        "default": true,
                        "description": "Skip auth",
                        "name": "X-Skip-Auth",
                        "in": "header"
                    },
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "description": "User ID",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.IDsInput"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    }
                }
            }
        },
        "/user/pass": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Set user password by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Set user password by ID",
                "parameters": [
                    {
                        "enum": [
                            true,
                            false
                        ],
                        "type": "boolean",
                        "default": true,
                        "description": "Skip auth",
                        "name": "X-Skip-Auth",
                        "in": "header"
                    },
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "User email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Password model",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.PasswordInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Reset user password by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Reset user password by ID",
                "parameters": [
                    {
                        "enum": [
                            true,
                            false
                        ],
                        "type": "boolean",
                        "default": true,
                        "description": "Skip auth",
                        "name": "X-Skip-Auth",
                        "in": "header"
                    },
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "User email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update user by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update user by ID",
                "parameters": [
                    {
                        "enum": [
                            true,
                            false
                        ],
                        "type": "boolean",
                        "default": true,
                        "description": "Skip auth",
                        "name": "X-Skip-Auth",
                        "in": "header"
                    },
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User model",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.UserInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.UserOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                },
                "object": {}
            }
        },
...
        "github_com_raulaguila_go-api_internal_core_dto.IDsInput": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "github_com_raulaguila_go-api_internal_core_dto.PaginatedOutput-github_com_raulaguila_go-api_internal_core_dto_UserOutput": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.UserOutput"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.PaginationOutput"
                }
            }
        },
        "github_com_raulaguila_go-api_internal_core_dto.PaginationOutput": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_raulaguila_go-api_internal_core_dto.PasswordInput": {
            "type": "object",
            "required": [
                "password",
                "password_confirm"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 6
                },
                "password_confirm": {
                    "type": "string"
                }
            }
        },
...
        "github_com_raulaguila_go-api_internal_core_dto.RoleOutput": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_raulaguila_go-api_internal_core_dto.UserInput": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 5
                },
                "role_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 5
                }
            }
        },
        "github_com_raulaguila_go-api_internal_core_dto.UserOutput": {
            "type": "object",
            "properties": {
                "corp_id": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "new": {
                    "type": "boolean"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.RoleOutput"
                    }
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and the JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}